+++ zapier/resources/table.js
@@ -1,5 +1,6 @@
 // get a list of tables
 const query = require('../tidb_client/PromiseClient')
+const request = require('../tidb_client/TiDBCloudClient')
 const performList = async (z, bundle) => {
   const host = bundle.inputData.host
   const port = bundle.inputData.port
@@ -16,7 +17,7 @@ const performList = async (z, bundle) => {
     `select tidb_table_id,table_name from information_schema.tables where table_schema='${bundle.inputData.database}'`,
   )
   if (error) {
-    throw new z.errors.Error('Execute SQL error', error, 400)
+    throw new z.errors.Error(`Execute SQL error: ${error}`)
   }
 
   const result = []
@@ -36,7 +37,7 @@ const performCreate = async (z, bundle) => {
 
   const [rows, error] = await query(host, user, port, tidbPassword, database, `${bundle.inputData.tableDDL}`)
   if (error) {
-    throw new z.errors.Error('Execute SQL error', error, 400)
+    throw new z.errors.Error(`Execute SQL error: ${error}`)
   }
 
   return Object.create(null)
@@ -59,7 +60,7 @@ const performSearch = async (z, bundle) => {
     `select table_name from information_schema.tables where table_schema='${bundle.inputData.database}' and table_name='${bundle.inputData.table}'`,
   )
   if (error) {
-    throw new z.errors.Error('Execute SQL error', error, 400)
+    throw new z.errors.Error(`Execute SQL error: ${error}`)
   }
   if (rows.length !== 0) {
     return [{ table: rows[0].table_name }]
@@ -71,14 +72,12 @@ const computedFields = async (z, bundle) => {
   if (bundle.inputData.clusterId === undefined) {
     return []
   }
-  const response = await z.request({
-    url:
-      `https://api.tidbcloud.com/api/v1beta/projects/${bundle.inputData.projectId}/clusters/${bundle.inputData.clusterId}`,
-    digest: {
-      username: bundle.authData.username,
-      password: bundle.authData.password,
-    },
-  })
+  const response = await request(
+    z,
+    `https://api.tidbcloud.com/api/v1beta/projects/${bundle.inputData.projectId}/clusters/${bundle.inputData.clusterId}`,
+    bundle.authData.username,
+    bundle.authData.password,
+  )
   const host = response.data.status.connection_strings.standard.host
   const port = response.data.status.connection_strings.standard.port
   const user = response.data.status.connection_strings.default_user